+++ src/assets/jss/material-dashboard-react/components/buttonStyle.jsx
@@ -40,7 +40,7 @@ const buttonStyle = {
       boxShadow:
         "0 14px 26px -12px rgba(153, 153, 153, 0.42), 0 4px 23px 0px rgba(0, 0, 0, 0.12), 0 8px 10px -5px rgba(153, 153, 153, 0.2)"
     },
-    "& .fab,& .fas,& .far,& .fal": {
+    "& .fab,& .fas,& .far,& .fal, &.material-icons": {
       position: "relative",
       display: "inline-block",
       top: "0",
@@ -60,7 +60,7 @@ const buttonStyle = {
       verticalAlign: "middle"
     },
     "&$justIcon": {
-      "& .fab,& .fas,& .far,& .fal": {
+      "& .fab,& .fas,& .far,& .fal,& .material-icons": {
         marginTop: "0px",
         position: "absolute",
         width: "100%",
@@ -219,7 +219,7 @@ const buttonStyle = {
     height: "41px",
     minWidth: "41px",
     width: "41px",
-    "& .fab,& .fas,& .far,& .fal,& svg": {
+    "& .fab,& .fas,& .far,& .fal,& svg,& .material-icons": {
       marginRight: "0px"
     },
     "&$lg": {
@@ -227,7 +227,7 @@ const buttonStyle = {
       minWidth: "57px",
       width: "57px",
       lineHeight: "56px",
-      "& .fab,& .fas,& .far,& .fal": {
+      "& .fab,& .fas,& .far,& .fal,& .material-icons": {
         fontSize: "32px",
         lineHeight: "56px"
       },
@@ -240,7 +240,7 @@ const buttonStyle = {
       height: "30px",
       minWidth: "30px",
       width: "30px",
-      "& .fab,& .fas,& .far,& .fal": {
+      "& .fab,& .fas,& .far,& .fal,& .material-icons": {
         fontSize: "17px",
         lineHeight: "29px"
       },