+++ src/assets/jss/material-dashboard-react/components/cardFooterStyle.jsx
@@ -28,7 +28,14 @@ const cardFooterStyle = {
       marginLeft: "3px",
       width: "16px",
       height: "16px"
-    }
+    },
+    "& .fab,& .fas,& .far,& .fal,& .material-icons": {
+      fontSize: "16px",
+      position: "relative",
+      top: "4px",
+      marginRight: "3px",
+      marginLeft: "3px",
+    },
   },
   cardFooterChart: {
     borderTop: "1px solid #eee"