+++ src/assets/jss/material-dashboard-react/components/cardHeaderStyle.jsx
@@ -37,12 +37,14 @@ const cardHeaderStyle = {
       height: "36px",
       margin: "10px 10px 4px"
     },
-    "&$cardHeaderStats i": {
+    "&$cardHeaderStats i,&$cardHeaderStats .material-icons": {
       fontSize: "36px",
       lineHeight: "56px",
       width: "56px",
       height: "56px",
-      textAlign: "center"
+      textAlign: "center",
+      overflow: "unset",
+      marginBottom: "1px"
     },
     "&$cardHeaderStats$cardHeaderIcon": {
       textAlign: "right"
@@ -65,7 +67,7 @@ const cardHeaderStyle = {
       background: "transparent",
       boxShadow: "none"
     },
-    "& i": {
+    "& i,& .material-icons": {
       width: "33px",
       height: "33px",
       textAlign: "center",