+++ src/assets/jss/material-dashboard-react/components/headerLinksStyle.jsx
@@ -41,6 +41,14 @@ const headerLinksStyle = theme => ({
         marginRight: "15px",
         marginLeft: "-15px"
       },
+      "& .fab,& .fas,& .far,& .fal,& .material-icons": {
+        fontSize: "24px",
+        lineHeight: "30px",
+        width: "24px",
+        height: "30px",
+        marginRight: "15px",
+        marginLeft: "-15px"
+      },
       "& > span": {
         justifyContent: "flex-start"
       }