+++ src/assets/jss/material-dashboard-react/tooltipStyle.jsx
@@ -2,9 +2,7 @@ const tooltipStyle = {
   tooltip: {
     padding: "10px 15px",
     minWidth: "130px",
-    color: "#555555",
     lineHeight: "1.7em",
-    background: "#FFFFFF",
     border: "none",
     borderRadius: "3px",
     boxShadow: