+++ src/views/Icons/Icons.jsx
@@ -2,10 +2,10 @@ import React from "react";
 import PropTypes from "prop-types";
 // @material-ui/core components
 import withStyles from "@material-ui/core/styles/withStyles";
-import Grid from "@material-ui/core/Grid";
 import Hidden from "@material-ui/core/Hidden";
 // core components
 import GridItem from "components/Grid/GridItem.jsx";
+import GridContainer from "components/Grid/GridContainer.jsx";
 import Card from "components/Card/Card.jsx";
 import CardHeader from "components/Card/CardHeader.jsx";
 import CardBody from "components/Card/CardBody.jsx";
@@ -15,7 +15,7 @@ import iconsStyle from "assets/jss/material-dashboard-react/views/iconsStyle.jsx
 function Icons(props) {
   const { classes } = props;
   return (
-    <Grid container>
+    <GridContainer>
       <GridItem xs={12} sm={12} md={12}>
         <Card plain>
           <CardHeader plain color="primary">
@@ -60,7 +60,7 @@ function Icons(props) {
           </CardBody>
         </Card>
       </GridItem>
-    </Grid>
+    </GridContainer>
   );
 }