+++ src/views/TableList/TableList.jsx
@@ -1,9 +1,9 @@
 import React from "react";
 // @material-ui/core components
 import withStyles from "@material-ui/core/styles/withStyles";
-import Grid from "@material-ui/core/Grid";
 // core components
 import GridItem from "components/Grid/GridItem.jsx";
+import GridContainer from "components/Grid/GridContainer.jsx";
 import Table from "components/Table/Table.jsx";
 import Card from "components/Card/Card.jsx";
 import CardHeader from "components/Card/CardHeader.jsx";
@@ -42,7 +42,7 @@ const styles = {
 function TableList(props) {
   const { classes } = props;
   return (
-    <Grid container>
+    <GridContainer>
       <GridItem xs={12} sm={12} md={12}>
         <Card>
           <CardHeader color="primary">
@@ -105,7 +105,7 @@ function TableList(props) {
           </CardBody>
         </Card>
       </GridItem>
-    </Grid>
+    </GridContainer>
   );
 }